requirements.txt
numpy
pandas
pyyaml
tqdm
seaborn
scipy
numba
PyQt5
plyer
